LayoutTests/inspector/canvas/setShaderProgramDisabled-expected.txt

Test disabling and re-enabling of shader programs.


== Running test suite: Canvas.setShaderProgramDisabled
-- Running test case: Canvas.setShaderProgramDisabled.disabledDrawArrays
PASS: Disabling a shader program should prevent it from drawing.

-- Running test case: Canvas.setShaderProgramDisabled.disabledDrawElements
PASS: Disabling a shader program should prevent it from drawing.

-- Running test case: Canvas.setShaderProgramDisabled.reenabledDrawArrays
PASS: Re-enabling a shader program should allow it to draw.

-- Running test case: Canvas.setShaderProgramDisabled.reenabledDrawElements
PASS: Re-enabling a shader program should allow it to draw.

-- Running test case: Canvas.setShaderProgramDisabled.invalidProgramId
PASS: Should produce an error.
Error: Missing program for given programId
